Moved clothespin to nipple folder
